requirements.txt
# Core ML
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
imbalanced-learn==0.12.2

# Anomaly Detection
pyod==2.0.1

# RAG / FAISS / Embeddings
faiss-cpu==1.8.0
sentence-transformers==3.0.1
anthropic>=0.40.0

# API
flask==3.0.3
flask-restx==1.3.0
flask-cors==4.0.1
gunicorn==22.0.0
marshmallow==3.21.3

# AWS
boto3==1.34.144

# Monitoring / Drift
evidently==0.4.30
mlflow==2.14.1
prometheus-client==0.20.0
scipy==1.13.1

# Explainability
shap==0.45.1

# Utilities
python-dotenv==1.0.1
pydantic==2.7.4
loguru==0.7.2
joblib==1.4.2
pyarrow>=10.0.0,<16
